refactor, docs: 3단계 피드백 리팩토링, 요구사항 리스트 수정
- 이번 주제에서 Step 관련 내용 없어 Enum step 제거
- 생성자에 차 대수, 시도 횟수 받는 로직으로 변경 (테스트 코드 포함 수정)
- 차 대수, 시도 횟수 관련 메소드 private로 변경
- readme 요구사항 리스트 수정
