+++ src/main/java/javaracingcar/controller/JavaRacingCar.java
@@ -1,8 +1,6 @@
-package javaracingcar;
+package javaracingcar.controller;
 
 import javaracingcar.entity.RacingCar;
-import javaracingcar.entity.RacingCarStepEnum;
-import javaracingcar.view.InputView;
 import javaracingcar.view.ResultView;
 
 import java.util.ArrayList;
@@ -18,24 +16,11 @@ public class JavaRacingCar {
     private final List<RacingCar> cars = new ArrayList<>();
 
     private int tryCount = DEFAULT_TRY_COUNT_VALUE;
-    private RacingCarStepEnum step = RacingCarStepEnum.INIT;
 
-    public JavaRacingCar() {
-        requireCarCount(InputView.requireCarCount());
-        requireTryCount(InputView.requireTryCount());
+    public JavaRacingCar(int carCount, int tryCount) {
+        requireCarCount(carCount);
+        requireTryCount(tryCount);
         start();
-    }
-
-    public void requireCarCount(int carCount) throws RuntimeException {
-        validateNegativeNumber(carCount);
-        initCars(carCount);
-        this.step = RacingCarStepEnum.CONFIRMED_CAR_COUNT;
-    }
-
-    public void requireTryCount(int tryCount) throws RuntimeException {
-        validateNegativeNumber(tryCount);
-        this.step = RacingCarStepEnum.CONFIRMED_TRY_COUNT;
-        this.tryCount = tryCount;
     }
 
     public List<RacingCar> start() {
@@ -48,6 +33,16 @@ public class JavaRacingCar {
 
     public int random() {
         return random.nextInt(RANDOM_VALUE_RANGE);
+    }
+
+    private void requireCarCount(int carCount) throws RuntimeException {
+        validateNegativeNumber(carCount);
+        initCars(carCount);
+    }
+
+    private void requireTryCount(int tryCount) throws RuntimeException {
+        validateNegativeNumber(tryCount);
+        this.tryCount = tryCount;
     }
 
     private void validateNegativeNumber(int count) {
@@ -77,10 +72,6 @@ public class JavaRacingCar {
 
     public int tryCount() {
         return this.tryCount;
-    }
-
-    public RacingCarStepEnum step() {
-        return this.step;
     }
 
     public List<RacingCar> cars() {
