--- src/main/java/javaracingcar/entity/RacingCarStepEnum.java
@@ -1,7 +1,0 @@
-package javaracingcar.entity;
-
-public enum RacingCarStepEnum {
-    INIT,
-    CONFIRMED_CAR_COUNT,
-    CONFIRMED_TRY_COUNT
-}
