+++ src/test/java/javaracingcar/README.md
@@ -8,6 +8,6 @@
 
 - Random 객체가 0 ~ 9 범위내의 값을 출력하는지 검증
 - Random 범위 내에서 값의 전진 또는 정지 인지 boolean 값 검증(예시 : 0 ~ 3 => false, 4 ~ 9 => true)
-- 객체 생성, 입력 후 변수 step 값 검증
 - 차 대수, 시도 횟수가 음수면 RuntimeException (예시 : -4)
 - 시도 횟수가 5이면 반복문으로 랜덤 5번 순회 후 전진 횟수가 0 ~ 5사이의 값인 지 검증
+- 생성 시 차 대수, 시도 횟수 검증
